server/requirements.txt
flask
flask-cors
torch
torchvision
numpy
pillow